Split the streaming example server so the API it demonstrates is findable
main.rs had grown to 445 lines, past this repo's own ~400-line split rule,
and buried the `render_streaming` call at line 251 between the protocol build
and the weather endpoint. A reader opening the example to learn the streaming
API found everything except the streaming API.

Each file now owns one concern, and `render_page` is the first function in
main.rs:

  main.rs   268  routes, `render_page`, the weather endpoint, CLI args
  pacing.rs 230  `CheckpointPacedWriter` and the flush schedule it consumes
  app.rs    191  protocol build and sample feed data
  assets.rs 172  cache-header policy for `dist/`
  jitter.rs 149  the demo's randomized delays

`paced_writer.rs` becomes `pacing.rs` and absorbs `gap_is_paced` plus its
constants and test, which were in main.rs despite existing only to drive the
writer in this file. `feed_batches` returns a `Vec` of batches instead of a
3-tuple, so adding one is a data edit rather than a signature change, and a
new test asserts it stays in step with the pacing schedule.

Behavior is unchanged: byte-identical response shape, 8/8 Playwright scenarios
pass, and the pacing schedule test still pins the flush-index-to-boundary
mapping.
