+++ examples/app/streaming/server/src/pacing.rs
@@ -1,7 +1,8 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-//! Example-only [`FlushWriter`] adapter that paces boundary flushes.
+//! Example-only [`FlushWriter`] adapter that paces boundary flushes, and the
+//! schedule this demo drives it with.
 //!
 //! `render_streaming` flushes synchronously at each committed boundary
 //! checkpoint. Left alone, a fast in-process render commits all of this
@@ -24,11 +25,34 @@
 //! The delay is supplied per flush rather than as one fixed duration,
 //! because this app's boundaries do not all deserve the same gap: the
 //! weather shell must hand straight over to the composer, while each feed
-//! batch should arrive after a visible, randomized pause.
+//! batch should arrive after a visible, randomized pause. [`gap_is_paced`]
+//! is that mapping.
 
 use std::time::Duration;
 
 use webui_handler::{FlushWriter, ResponseWriter, Result};
+
+/// Zero-based index of the flush whose pause precedes feed batch 1.
+///
+/// Flush 0 commits the weather shell, which carries no server data — the
+/// composer must follow it immediately, so that gap is never paced. Flush 1
+/// commits the composer, and its pause is the wait before the first feed
+/// batch arrives.
+const FIRST_FEED_GAP: usize = 1;
+
+/// Feed batches, and therefore jittered gaps: one before each batch.
+pub(crate) const FEED_BATCH_COUNT: usize = 3;
+
+/// Whether the pause *after* flush `index` should be paced.
+///
+/// Flushes [`FIRST_FEED_GAP`] onward commit the composer and the first two
+/// feed batches, and each of their pauses is the wait before the next feed
+/// batch. Everything after that — the last batch, the implicit tail
+/// checkpoint, and the terminal record — closes the response without
+/// further delay.
+pub(crate) fn gap_is_paced(index: usize) -> bool {
+    (FIRST_FEED_GAP..FIRST_FEED_GAP + FEED_BATCH_COUNT).contains(&index)
+}
 
 /// Wraps a [`FlushWriter`] and sleeps for a caller-chosen duration after
 /// each flush.
@@ -80,7 +104,7 @@ impl<W: FlushWriter, D: FnMut(usize) -> Duration> FlushWriter for CheckpointPace
 
 #[cfg(test)]
 mod tests {
-    use super::CheckpointPacedWriter;
+    use super::{gap_is_paced, CheckpointPacedWriter};
     use std::time::{Duration, Instant};
     use webui_handler::{FlushWriter, HandlerError, ResponseWriter, Result};
 
@@ -211,15 +235,20 @@ mod tests {
         assert!(writer.inner.ended);
     }
 
+    /// The exact schedule `render_page` installs, so the flush-index-to-
+    /// boundary mapping is asserted rather than implied by a comment.
     #[test]
-    fn an_all_zero_schedule_never_sleeps() {
-        let mut writer = CheckpointPacedWriter::new(RecordingWriter::default(), |_| Duration::ZERO);
-        let start = Instant::now();
-        for _ in 0..5 {
-            writer
-                .flush()
-                .unwrap_or_else(|e| panic!("flush failed: {e}"));
-        }
-        assert!(start.elapsed() < Duration::from_millis(50));
+    fn only_the_gaps_before_feed_batches_are_paced() {
+        // Flush 0 commits the weather shell; the composer must follow it
+        // immediately or the highest-priority island is delayed behind a
+        // boundary that carries no server data at all.
+        assert!(!gap_is_paced(0), "the weather-to-composer gap must be free");
+        // Flushes 1, 2 and 3 precede feed batches 1, 2 and 3.
+        assert!(gap_is_paced(1));
+        assert!(gap_is_paced(2));
+        assert!(gap_is_paced(3));
+        // The tail checkpoint and terminal record close the response.
+        assert!(!gap_is_paced(4), "the response must close promptly");
+        assert!(!gap_is_paced(5));
     }
 }
